REVIEW: incorporated ui comments
